APP_API/requirements.txt

djangorestframework
